requirements.txt
# Note: this is the list of dependencies that should be installed for the project.
# It is recommended to install these in a Miniconda/Anaconda environment with Python 3.10.

black  # code formatting
deepdiff  # generic data manipulation utility
deeplake  # i/o library for activeloop datasets
einops  # to simplify some tensor ops
flake8  # code formatting
flake8-docstrings  # code formatting
gitpython  # for git i/o and hash tagging
h5py  # i/o library for HDF5 datasets
hydra-colorlog==1.2.0  # configuration/logging
hydra-core==1.3.0  # configuration/logging
imagesize  # generic imgproc utility
isort  # code formatting
jupyter  # for data exploration notebooks
lightning-bolts  # for extra lightning goodies
lz4  # data compression
matplotlib  # visualization/logging
mlflow  # visualization/logging
mock  # testingire
nbstripout  # code formatting
numpy  # generic data processing utility
opencv-python  # generic image processing utility
pre-commit # code formatting and static analysis
pyrootutils # tooling for python proj
pytest  # testing
python-dotenv  # loads env variables from .env
pytorch-lightning==1.8.3  # training helper
pyyaml  # configuration/logging
rich  # better text formatting
segmentation-models-pytorch  # model architectures
setuptools  # package install tool
torch>=1.10.0  # deep learning backbone
torchmetrics>=0.11.0  # model evaluation
torchvision>=0.11.0  # model architectures + data utilities
tqdm  # logging and progress bars
wandb  # visualization/logging
